feat: create a package for the local api gateway websocket
