+++ package.json
@@ -18,7 +18,7 @@
     "coverage": "lerna run --stream coverage --concurrency 1 -- --color",
     "eslint": "lerna run --parallel eslint -- --color",
     "lint": "run-p eslint stylelint",
-    "start:apigateway-ws": "nodemon --exec 'babel-node -x .js,.ts' scripts/local/start-api-gateway-websocket.ts",
+    "start:apigateway-ws": "lerna run --stream --scope local-api-gateway-websocket start -- --color",
     "start:client": "lerna run --stream --scope client start -- --color",
     "stylelint": "lerna run --parallel stylelint -- --color",
     "tsc": "lerna run --parallel tsc -- --pretty",
@@ -26,8 +26,8 @@
   },
   "devDependencies": {
     "@babel/cli": "^7.8.4",
-    "@babel/core": "^7.8.4",
-    "@babel/node": "^7.8.4",
+    "@babel/core": "^7.8.7",
+    "@babel/node": "^7.8.7",
     "@babel/plugin-proposal-class-properties": "^7.8.3",
     "@babel/plugin-proposal-decorators": "^7.8.3",
     "@babel/preset-env": "^7.8.4",
@@ -35,7 +35,6 @@
     "@babel/preset-typescript": "^7.8.3",
     "@typescript-eslint/eslint-plugin": "^2.20.0",
     "@typescript-eslint/parser": "^2.20.0",
-    "aws-sdk": "^2.632.0",
     "babel-eslint": "^10.0.3",
     "babel-jest": "25.1.0",
     "babel-plugin-module-resolver": "^4.0.0",
@@ -51,10 +50,8 @@
     "faker": "^4.1.0",
     "jest": "25.1.0",
     "lerna": "^3.20.2",
-    "nodemon": "^2.0.2",
     "npm-run-all": "^4.1.5",
     "ts-node": "^8.6.2",
-    "typescript": "^3.8.2",
-    "ws": "^7.2.1"
+    "typescript": "^3.8.2"
   }
 }
