+++ packages/client/package.json
@@ -31,7 +31,9 @@
     "mobx-react": "^6.1.7",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
-    "uuid": "^7.0.0"
+    "shared": "./packages/shared",
+    "uuid": "^7.0.0",
+    "ws": "^7.2.1"
   },
   "devDependencies": {
     "@babel/plugin-proposal-class-properties": "^7.8.3",
