+++ packages/local-api-gateway-websocket/.dockerignore
@@ -1,0 +1,1 @@
+node_modules
