+++ packages/local-api-gateway-websocket/babel.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+    ignore: ['node_modules'],
+    presets: [
+        '@babel/preset-typescript',
+        [
+            '@babel/preset-env',
+            {
+                targets: { node: 'current' }
+            }
+        ]
+    ]
+};
