+++ scripts/local/build-env.sh
@@ -17,5 +17,3 @@ aws --endpoint-url=http://localstack:4574 lambda create-function --function-name
     --handler lambda-entry.websocketDisconnection \
     --runtime nodejs12.x \
     --role lambda_role
-
-yarn run start:apigateway-ws
